Generate python bindings for mean and nanmean
