+++ lib/dataset/include/scipp/dataset/mean.h
@@ -4,6 +4,11 @@
 /// @author Simon Heybrock
 #pragma once
 
+/*
+ * These functions are not generated by CMake because they require
+ * slightly different operations from other reductions.
+ */
+
 #include "scipp/dataset/dataset.h"
 
 namespace scipp::dataset {
@@ -12,10 +17,5 @@ SCIPP_DATASET_EXPORT DataArray mean(const DataArray &a, const Dim dim);
 SCIPP_DATASET_EXPORT DataArray mean(const DataArray &a);
 SCIPP_DATASET_EXPORT Dataset mean(const Dataset &d, const Dim dim);
 SCIPP_DATASET_EXPORT Dataset mean(const Dataset &d);
-
-SCIPP_DATASET_EXPORT DataArray nanmean(const DataArray &a, const Dim dim);
-SCIPP_DATASET_EXPORT DataArray nanmean(const DataArray &a);
-SCIPP_DATASET_EXPORT Dataset nanmean(const Dataset &d, const Dim dim);
-SCIPP_DATASET_EXPORT Dataset nanmean(const Dataset &d);
 
 } // namespace scipp::dataset
