+++ lib/dataset/test/dataset_operations_test.cpp
@@ -5,8 +5,8 @@
 
 #include "dataset_test_common.h"
 #include "scipp/dataset/dataset.h"
+#include "scipp/dataset/mean.h"
 #include "scipp/dataset/nansum.h"
-#include "scipp/dataset/reduction.h"
 #include "scipp/dataset/sum.h"
 #include "test_macros.h"
