+++ lib/dataset/test/event_data_operations_consistency_test.cpp
@@ -4,7 +4,7 @@
 
 #include "scipp/dataset/bins.h"
 #include "scipp/dataset/histogram.h"
-#include "scipp/dataset/reduction.h"
+#include "scipp/dataset/mean.h"
 #include "scipp/dataset/sum.h"
 
 using namespace scipp;
